bugfix: fix functions to crawl syllabus and send to database
